fix: retire le cast du type integer pour le projet structurant code (000)

+++ data_management/data_factory/models/exposition/indicateur_projet_structurant.sql
@@ -2,7 +2,7 @@ SELECT
     DISTINCT ON (indicateur_nom)
     {{ dbt_utils.surrogate_key(['indicateur_nom']) }} as id,
     data_kpis.indicateur_nom as nom,
-    data_kpis.projet_structurant_code::INTEGER,
+    data_kpis.projet_structurant_code,
     CASE
         WHEN type_nom = 'RealisationHausse' OR type_nom = 'RealisationBaisse' THEN 'DEPL'
         WHEN type_nom = 'FinancierHausse' OR type_nom = 'FinancierBaisse' THEN 'FINANCIER'
